+++ checkout-pricing/src/main/java/com/tdd/PricingRules.java
@@ -8,12 +8,12 @@ import com.tdd.rules.cross.CrossSkuRule;
 import java.util.*;
 
 public class PricingRules {
-    private Map<String, Integer> unitPrices = new HashMap<>();
+    private Map<String, Double> unitPrices = new HashMap<>();
     private Map<String, List<PricingOption>> options = new HashMap<>();
     private List<SkuDiscount> skuDiscounts = new ArrayList<>();
     private final List<CrossSkuRule> crossSku = new ArrayList<>();
 
-    public PricingRules(Map<String, Integer> unitPrices, Map<String, List<PricingOption>> options,
+    public PricingRules(Map<String, Double> unitPrices, Map<String, List<PricingOption>> options,
                         List<CrossSkuBuyXGetYFree> freeRules, List<CrossSkuBuyXGetYDiscount> discountRules,
                         List<SkuDiscount> skuDiscounts) {
         this.unitPrices = unitPrices;
@@ -25,19 +25,19 @@ public class PricingRules {
 
     public PricingRules() {}
 
-    public PricingRules(Map<String, Integer> unitPrices) {
+    public PricingRules(Map<String, Double> unitPrices) {
         this.unitPrices = unitPrices;
     }
 
-    public void addUnitPrice(String sku, int price) {
+    public void addUnitPrice(String sku, double price) {
         unitPrices.put(sku, price);
     }
 
-    public int getUnitPrice(String sku) {
+    public double getUnitPrice(String sku) {
         return unitPrices.get(sku);
     }
 
-    public void addSpecialPrice(String sku, int quantity, int price, int priority, boolean stackable) {
+    public void addSpecialPrice(String sku, int quantity, double price, int priority, boolean stackable) {
         options.computeIfAbsent(sku, _ -> new ArrayList<>())
                 .add(new SpecialPrice(quantity, price, priority, stackable));
     }
@@ -50,15 +50,15 @@ public class PricingRules {
 
     public void addBuyXGetYFree(String sku, int buy, int free, int priority, boolean stackable) {
         int quantity = BuyXGetYFree.calculateQuantity(buy, free);
-        int price =  BuyXGetYFree.calculatePrice(buy, getUnitPrice(sku));
+        double price =  BuyXGetYFree.calculatePrice(buy, getUnitPrice(sku));
 
         options.computeIfAbsent(sku, _ -> new ArrayList<>())
                 .add(new BuyXGetYFree(quantity, price, priority, stackable));
     }
 
     public void addBuyXGetYDiscount(String sku, int buy, int get, double discount, int priority, boolean stackable) {
-        int unitPrice = getUnitPrice(sku);
-        int price = BuyXGetYDiscount.calculatePrice(buy, unitPrice, get, discount);
+        double unitPrice = getUnitPrice(sku);
+        double price = BuyXGetYDiscount.calculatePrice(buy, unitPrice, get, discount);
         int quantity = BuyXGetYDiscount.calculateQuantity(buy, get);
 
         options.computeIfAbsent(sku, _ -> new ArrayList<>())
