+++ checkout-pricing/src/main/java/com/tdd/api/samples/Ruleset.java
@@ -6,11 +6,11 @@ import java.util.Map;
 import static java.util.stream.Collectors.toMap;
 
 public interface Ruleset {
-    Map<String, Integer> DEFAULT_UNIT_PRICES = Arrays.stream(SKUs.values())
+    Map<String, Double> DEFAULT_UNIT_PRICES = Arrays.stream(SKUs.values())
             .collect(toMap(Enum::name, sku -> sku.unitPrice));
 
-    static Map<String, Integer> defaultUnitPricesWithChange(String sku, int unitPrice) {
-        Map<String, Integer> unitPrices = DEFAULT_UNIT_PRICES.entrySet().stream()
+    static Map<String, Double> defaultUnitPricesWithChange(String sku, double unitPrice) {
+        Map<String, Double> unitPrices = DEFAULT_UNIT_PRICES.entrySet().stream()
                 .collect(toMap(Map.Entry::getKey, Map.Entry::getValue));
         unitPrices.put(sku, unitPrice);
