+++ checkout-pricing/src/main/java/com/tdd/rules/BuyXGetYDiscount.java
@@ -1,12 +1,12 @@
 package com.tdd.rules;
 
-public record BuyXGetYDiscount(int quantity, int price, int priority, boolean stackable) implements PricingOption {
+public record BuyXGetYDiscount(int quantity, double price, int priority, boolean stackable) implements PricingOption {
 
     public static int calculateQuantity(int buy, int get) {
         return buy + get;
     }
 
-    public static int calculatePrice(int buy, int unitPrice, int get, double discount) {
-        return (int)(buy * unitPrice + get * unitPrice * (1-discount));
+    public static double calculatePrice(int buy, double unitPrice, int get, double discount) {
+        return buy * unitPrice + get * unitPrice * (1-discount);
     }
 }
