+++ checkout-pricing/src/main/java/com/tdd/rules/BuyXGetYFree.java
@@ -1,12 +1,12 @@
 package com.tdd.rules;
 
-public record BuyXGetYFree(int quantity, int price, int priority, boolean stackable) implements PricingOption {
+public record BuyXGetYFree(int quantity, double price, int priority, boolean stackable) implements PricingOption {
 
     public static int calculateQuantity(int buy, int get) {
         return buy + get;
     }
 
-    public static int calculatePrice(int buy, int unitPrice) {
+    public static double calculatePrice(int buy, double unitPrice) {
         return buy * unitPrice;
     }
 }
