+++ checkout-pricing/src/main/java/com/tdd/rules/SpecialPrice.java
@@ -1,4 +1,4 @@
 package com.tdd.rules;
 
-public record SpecialPrice(int quantity, int price, int priority, boolean stackable) implements PricingOption {
+public record SpecialPrice(int quantity, double price, int priority, boolean stackable) implements PricingOption {
 }
